+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Gunicorn HTTP Request/Response Smuggling vulnerability by updating gunicorn (22 -> 23.0.0)
 - Jinja2 vulnerabilities by updating jinja2 (3.1.4 -> 3.1.6)
 - Missing images on loading intervals documentation file
+- Missing file error when d4 file is provided as an HTTP resource to the coverage endpoints `/coverage/d4/interval` and `/coverage/d4/interval_file` and /coverage/samples/predicted_sex
 
 ## [3.4]
 ### Changed